Added support for PatchOperationRemove on RecipeDefs
Added support for PatchOperationRemove on RecipeDefs.

PatchOperationReplace doesn't work on RecipeDefs automatically added by this mod.
To replace a value you need to completely remove the recipe and add a new one with the needed values.
As base can the RecipeDef/BaseRecipe.xml be used with the ParentName="MinesAutomated_RecipeDef_Base" attribute.